Workaround(?) windows binding error
For some reason the code this replaced worked fine on linux/macOS
but failed with a rebinding error on Windows. I'm not sure if that
error is legit and missing from non-Windows, or if its erronious.
Regardless this changes the assignment to a binding to workaround
the binding issue in `Distribution::Path.TWEAK`s setting of
`self.meta<files> := my %files`